fixed bug which lead to wrong outliers
Behaviour before: if less than the set minNumberOfLines were found the point was marked as outliers.
Behaviour now: if less than the set minNumberOfLines were found the areaSize is increased until it reaches the maximum. Finally, the point is marked as an outlier only if no lines were found within the areasize